requirements.txt
fastapi
uvicorn
faiss-cpu
together
numpy
pydantic
